requirements.txt
PyQt
twisted
qt4reactor
# requests is used in tests
requests
